add test_utils file for shared helpers
